Use /proc/litmus/domains to set up affinity masks.
This patch replaces the algorithm used to compute
CPU affinity masks. The old algorithm was fragile
and would break on systems with multiple levels
of shared caches. (As found on some older Intel
chips, and as one would probably find on a system
with hyper-threading enabled.) The bug stems from
fact that Linux enumerates consecutive CPUs in a way
that distributes them across the physical system,
thereby reducing contention resources (e.g., cache).